src/requirements.txt
pandas
openpyxl
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
sentence-transformers
python-dotenv
langchain-chroma